refactor : relocation
디렉토리 정리 및 파일 세분화
