+++ app/src/main/java/org/android/go/sopt/login/MyPageActivity.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt
+package org.android.go.sopt.login
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
