+++ app/src/main/java/org/android/go/sopt/playlist/GalleryFragment.kt
@@ -1,23 +1,23 @@
-package org.android.go.sopt
+package org.android.go.sopt.playlist
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import org.android.go.sopt.databinding.FragmentSearchBinding
+import org.android.go.sopt.databinding.FragmentGalleryBinding
 
-class SearchFragment : Fragment() {
-    private var _binding: FragmentSearchBinding? = null
-    private val binding: FragmentSearchBinding
-        get() = requireNotNull(_binding) { "${this::class.java.simpleName}에서 에러가 발생했습니다."}
+class GalleryFragment : Fragment() {
+    private var _binding: FragmentGalleryBinding? = null
+    private val binding: FragmentGalleryBinding
+        get() = requireNotNull(_binding) {  "${this::class.java.simpleName}에서 에러가 발생했습니다." }
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentSearchBinding.inflate(inflater, container, false)
+        _binding = FragmentGalleryBinding.inflate(inflater, container, false)
         return binding.root
     }
