+++ app/src/main/java/org/android/go/sopt/playlist/PlaylistSong.kt
@@ -1,0 +1,9 @@
+package org.android.go.sopt.playlist
+
+import androidx.annotation.DrawableRes
+
+data class PlaylistSong(
+    @DrawableRes val album: Int,
+    val artist: String,
+    val title: String
+)
