+++ app/src/main/java/org/android/go/sopt/playlist/PlaylistViewHolder.kt
@@ -1,0 +1,13 @@
+package org.android.go.sopt.playlist
+
+import androidx.recyclerview.widget.RecyclerView
+import org.android.go.sopt.databinding.ItemPlaylistSongBinding
+
+class PlaylistViewHolder (private val binding: ItemPlaylistSongBinding) :
+    RecyclerView.ViewHolder(binding.root) {
+    fun onBind(item: PlaylistSong) {
+        binding.ivAlbum.setImageDrawable(binding.root.context.getDrawable(item.album))
+        binding.tvArtist.text = item.artist
+        binding.tvTitle.text = item.title
+    }
+}
