+++ app/src/main/java/org/android/go/sopt/playlist/mockPlayList.kt
@@ -1,0 +1,16 @@
+package org.android.go.sopt.playlist
+
+import org.android.go.sopt.R
+
+val mockPlayList = listOf<PlaylistSong>(
+    PlaylistSong(
+        album = R.drawable.ic_launcher_background,
+        artist = "Lacuna",
+        title = "Far Away",
+    ),
+    PlaylistSong(
+        album = R.drawable.ic_launcher_background,
+        artist = "최유리",
+        title = "동네",
+    )
+)
