+++ app/src/main/res/layout/activity_my_page.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MyPageActivity">
+    tools:context=".login.MyPageActivity">
 
     <ImageView
         android:id="@+id/my_page_image"
