+++ app/src/main/res/layout/fragment_home.xml
@@ -15,7 +15,7 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        tools:listitem="@layout/item_song" />
+        tools:listitem="@layout/item_playlist_song" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
