+++ app/src/main/res/menu/menu.xml
@@ -3,13 +3,13 @@
     <item
         android:id="@+id/menu_home"
         android:icon="@drawable/ic_home"
-        android:title="홈" />
+        android:title="@string/menu_item_home" />
     <item
         android:id="@+id/menu_search"
-        android:icon="@drawable/ic_home"
-        android:title="검색" />
+        android:icon="@drawable/ic_search"
+        android:title="@string/menu_item_search" />
     <item
         android:id="@+id/menu_gallery"
-        android:icon="@drawable/ic_home"
-        android:title="갤러리" />
+        android:icon="@drawable/ic_gallery"
+        android:title="@string/menu_item_gallery" />
 </menu>
